dom/nodes/insertion-effects/Node-appendChild-script-and-stylesheet-link.tentative.html

<!doctype html>
<meta charset=utf-8>
<title>Node.appendChild: inserting script and stylesheet link</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<body>
<script>
test(() => {
  window.happened = [];
  window.link = document.createElement("link");
  link.rel = "stylesheet";
  link.href = "data:text/css,";

  const script = document.createElement("script");
  script.textContent = `
    happened.push(link.sheet ? "sheet" : null);
  `;

  const df = document.createDocumentFragment();
  df.appendChild(script);
  df.appendChild(link);

  assert_array_equals(happened, []);
  document.body.appendChild(df);
  assert_array_equals(happened, ["sheet"]);
}, "Inserting in a fragment");

test(() => {
  window.happened = [];
  window.link = document.createElement("link");
  link.rel = "stylesheet";
  link.href = "data:text/css,";

  const script = document.createElement("script");
  script.textContent = `
    happened.push(link.sheet ? "sheet" : null);
`;

  const div = document.createElement("div");
  div.appendChild(script);
  div.appendChild(link);

  assert_array_equals(happened, []);
  document.body.appendChild(div);
  assert_array_equals(happened, ["sheet"]);
}, "Inserting in a div");


test(() => {
  window.happened = [];
  window.link = document.createElement("link");
  link.rel = "stylesheet";
  link.href = "data:text/css,";

  const script = document.createElement("script");
  script.textContent = `
    happened.push(link.sheet ? "sheet" : null);
`;

  assert_array_equals(happened, []);
  document.body.append(script, link);
  assert_array_equals(happened, ["sheet"]);
}, "Inserting in an append");
</script>
